Move model logic for `category` and `keyword` under `models`
