+++ src/lib.rs
@@ -65,7 +65,6 @@ use util::{R404, C, R};
 pub mod app;
 pub mod badge;
 pub mod boot;
-pub mod category;
 pub mod config;
 pub mod crate_owner_invitation;
 pub mod db;
@@ -75,7 +74,6 @@ pub mod download;
 pub mod git;
 pub mod github;
 pub mod http;
-pub mod keyword;
 pub mod krate;
 pub mod owner;
 pub mod render;
