+++ src/models/mod.rs
@@ -1,11 +1,14 @@
 pub use badge::{Badge, MaintenanceStatus};
-pub use category::{Category, CrateCategory, NewCategory};
+pub use self::category::{Category, CrateCategory, NewCategory};
 pub use crate_owner_invitation::NewCrateOwnerInvitation;
 pub use dependency::{Dependency, Kind, ReverseDependency};
 pub use download::VersionDownload;
-pub use keyword::{CrateKeyword, Keyword};
+pub use self::keyword::{CrateKeyword, Keyword};
 pub use krate::{Crate, CrateDownload, Follow, NewCrate};
 pub use owner::{CrateOwner, NewTeam, Owner, OwnerKind, Rights, Team};
 pub use user::{Email, NewUser, User};
 pub use token::ApiToken;
 pub use version::{NewVersion, Version};
+
+mod category;
+mod keyword;
